requirements.txt
flask==3.1.2
waitress==3.0.2
python-dotenv==1.2.1